close #271 (#272)
* remove spurious 64 in type signature; 
* reduce testing demands; 
* add 32-bit tests to travis
* adjust test suite for baked in 64-bit expectations

+++ src/polynomials/LaurentPolynomial.jl
@@ -68,8 +68,8 @@ LaurentPolynomial(3.0 + 2.0*x + 1.0*x²)
 struct LaurentPolynomial{T <: Number} <: StandardBasisPolynomial{T}
     coeffs::Vector{T}
     var::Symbol
-    m::Base.RefValue{Int64}
-    n::Base.RefValue{Int64}
+    m::Base.RefValue{Int}
+    n::Base.RefValue{Int}
     function LaurentPolynomial{T}(coeffs::AbstractVector{T},
                                   m::Int,
                                   var::Symbol=:x) where {T <: Number}
@@ -126,7 +126,7 @@ end
 
 ## Alternate with range specified
 function  LaurentPolynomial{T}(coeffs::AbstractVector{S},
-                               rng::UnitRange{Int64},
+                               rng::UnitRange{Int},
                                var::Symbol=:x) where {T <: Number, S <: Number}
     LaurentPolynomial{T}(T.(coeffs), first(rng), var)
 end
